AO3: fix fic-identity hash mismatch + finish X4 Pro touch input
Root-cause fix: the AO3 library identified fics with std::hash<std::string>
(32-bit), but CrossInk's own epub cache migrated to ZipFile::fnvHash64
(64-bit) for its cache directory names a while back. Every AO3 sidecar
(ao3_library_info, holding the real file path plus all rich metadata) was
being written and read under the OLD hash scheme, but that no longer
matched where the shared epub cache actually put things once a book was
opened normally -- so on a from-scratch flash, metadata came back blank
(only the ~11-char index-fallback title showed) and tapping/selecting a fic
had no file path to open. xAO3 was unaffected because its CrossPoint base
predates the fnvHash64 migration. This was invisible until this session
because the AO3 port had never been exercised on real hardware.

Fix: switch fic-identity everywhere to ZipFile::fnvHash64 (matching
Epub::cachePathForFilePath), widening CompactIndexRecord/ViewEntry's
cacheHash from uint32_t to uint64_t. This changes the on-disk index record
size (241 -> 245 bytes) and hash values, so the index format version is
bumped 2 -> 3 (all read-side checks updated to match, including two in
Ao3LibraryActivity.cpp and Ao3EndOfBookSeriesActivity.cpp that were
initially missed and caused a false-negative "library index missing or
corrupt" on an otherwise-valid freshly-rebuilt index). An old v2 index is
rejected outright rather than misread; re-running Index New Books rebuilds
it cleanly.

Also fixes an indexing-loop bug found while rebuilding a large (500+ book)
library on-device: DIR_INDEXING only remembered the first 10 failed books
for re-processing exclusion, so once more than 10 books failed to scrape,
the excess were endlessly re-collected every batch and the progress
counter ran past the total without ever reaching completion. Failed
hashes are now tracked in a separate uncapped vector for exclusion;
failedBooks stays capped at 10 for the failed-list UI only.

Touch input for the X4 Pro (CROSSINK_APP_CAP_TOUCH), completing what
BookFusion already had and AO3 (ported button-only from xAO3) didn't:
Ao3LibraryActivity (library list tap-to-open/swipe-page/filter-tap, manage
panel, filter panel, fandom/relationship pickers), Ao3IndexActivity
(tap-to-confirm through the indexing prompts), Ao3LibrarySettingsActivity
(tap a row, cleanup prompts), Ao3FolderPickerActivity (tap to select,
long-press to navigate into a folder, matching the existing short/hold-
Confirm semantics). Every touch query used is a constexpr no-op when
CROSSINK_APP_CAP_TOUCH=0 (verified via the build's own
check_app_touch_gate.py audit), so default/sticky remain button-only.

Row hit-testing is done via TouchRegistry (registered at render time,
read back with wasItemTapped/wasScreenTapped), the same mechanism
BookFusion's screen.list relies on -- not hand-computed row geometry,
which proved unreliable on-device. In-place actions route through a
local per-frame flag rather than injectRelease(), which the main loop
only clears on the button-shortcut dispatch path, not the normal
activity loop -- an injected release there would never clear and would
fire every frame forever (this was the cause of the filter/manage panel
infinite-cycling bug found on-device).

Hardware-verified on X4 Pro: full library re-index (500+ books)
completes, metadata populates, tapping a fic opens it, filter/manage
panels and folder picker work via touch, buttons still work unchanged.

+++ src/activities/reader/Ao3EndOfBookSeriesActivity.h
@@ -12,7 +12,7 @@
 class Ao3EndOfBookSeriesActivity final : public Activity {
  public:
   Ao3EndOfBookSeriesActivity(GfxRenderer& renderer, MappedInputManager& mappedInput, std::string seriesName,
-                             uint32_t seriesHash, uint32_t originCacheHash, std::string originEpubPath)
+                             uint32_t seriesHash, uint64_t originCacheHash, std::string originEpubPath)
       : Activity("Ao3EndOfBookSeries", renderer, mappedInput),
         seriesName_(std::move(seriesName)),
         seriesHash_(seriesHash),
@@ -26,7 +26,7 @@ class Ao3EndOfBookSeriesActivity final : public Activity {
  private:
   std::string seriesName_;
   uint32_t seriesHash_;
-  uint32_t originCacheHash_;
+  uint64_t originCacheHash_;
   std::string originEpubPath_;
 
   std::vector<ViewEntry> viewEntries;
@@ -41,7 +41,7 @@ class Ao3EndOfBookSeriesActivity final : public Activity {
 
   void loadViewEntries();
   void loadPageCache(int page);
-  BookStatus getBookStatus(uint32_t cacheHash);
+  BookStatus getBookStatus(uint64_t cacheHash);
 
   void renderEntry(RenderLock& lock, int y, const ViewEntry& ve, int cacheSlot, bool selected);
   void drawAo3Square(RenderLock& lock, int x, int y, int s, char rating, char warning, bool completed,
